fix: :truck: Moved collision process to movement
Issue: #39
