+++ src/player.rs
@@ -21,10 +21,6 @@ impl Plugin for PlayerPlugin {
     fn build(&self, app: &mut App) {
         app
             .add_systems(Update, (
-                crate::wall::check_for_wall,
-                crate::block::collision::check_for_collision,
-            ).chain().run_if(in_state(AppState::Ingame)))
-            .add_systems(Update, (
                 block_rotation,
                 crate::block::rotation::rotation,
             ).chain().run_if(in_state(AppState::Ingame)))
